% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instaeda.R
\name{plot_basic_distributions}
\alias{plot_basic_distributions}
\title{Plot basic distributions for numeric and string features from a given dataframe}
\usage{
plot_basic_distributions(
  df,
  cols = NULL,
  include = NULL,
  colour_palette = "PuOr"
)
}
\arguments{
\item{df}{Dataframe from which to generate plots for each column from}

\item{cols}{List of columns to generate plots for. By default, NULL (perform on all numeric).}

\item{include}{select the data type to include. Supported values include NULL, "string" and "number". When selecting NULL, both string and number datatypes will be returned. By default, NULL.}

\item{colour_palette}{one of ggplot accepted colour schemes for the plots}
}
\value{
named list of ggplot objects with each name referencing a column name
}
\description{
Plot basic distributions for numeric and string features from a given dataframe
}
\examples{
 \dontrun{
 plot_basic_distributions(example_dataframe)
 }
}
