% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instaeda.R
\name{plot_info}
\alias{plot_info}
\title{Plot summary introduction for input data.}
\usage{
plot_info(data, title = "", color_config = list())
}
\arguments{
\item{data}{input data}

\item{title}{plot title}

\item{color_config}{configurations to manually change colors}
}
\value{
the ggplot object
}
\description{
Plot summary introduction for input data.
}
\examples{
\dontrun{
plot_info(example_dataframe)
}
}
