% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instaeda.R
\name{divide_and_fill}
\alias{divide_and_fill}
\title{Takes a dataframe, subsets selected columns and imputes missing values and returns a data frame.}
\usage{
divide_and_fill(
  dataframe,
  cols = NULL,
  strategy = "mean",
  random = FALSE,
  verbose = 0L
)
}
\arguments{
\item{dataframe}{Dataframe from which to take columns and
check for missing values.}

\item{cols}{List of columns to perform imputation on.
By default, NULL(perform on all numeric columns).}

\item{strategy}{string. imputation strategy, one of:
{'mean', 'median', 'random'}. By default, 'mean'.}

\item{random}{boolean, optional
When random == True, shuffles data frame before filling. By default, False.}

\item{verbose}{integer, optional
Controls the verbosity of the divide and fill. By default, 0.}
}
\value{
data.frame Data frame obtained after divide and fill on the
corresponding columns.
}
\description{
Takes a dataframe, subsets selected columns and imputes missing values and returns a data frame.
}
\examples{
\dontrun{
divide_and_fill(example_dataframe)
}
}
